added formatPercent and formatAmount methods in useNumberFormat hook
