+++ src/modules/pools/staking/StakingCard.jsx
@@ -18,7 +18,6 @@ import { useTokenSymbol } from "@/src/hooks/useTokenSymbol";
 import { config } from "@neptunemutual/sdk";
 import { useNetwork } from "@/src/context/Network";
 import { explainInterval } from "@/utils/formatter/interval";
-import { formatPercent } from "@/utils/formatter/percent";
 import { Badge } from "@/common/Badge/Badge";
 import { PoolTypes } from "@/src/config/constants";
 import { getApr } from "@/src/services/protocol/staking-pool/info/apr";
@@ -34,7 +33,7 @@ export const StakingCard = ({ data, tvl, getPriceByAddress }) => {
     key: data.key,
     type: PoolTypes.TOKEN,
   });
-  const { formatCurrency } = useNumberFormat();
+  const { formatCurrency, formatPercent } = useNumberFormat();
 
   const rewardTokenAddress = info.rewardToken;
   const stakingTokenSymbol = useTokenSymbol(info.stakingToken);
