+++ src/modules/reporting/active/ActiveReportSummary.jsx
@@ -10,7 +10,6 @@ import { truncateAddress } from "@/utils/address";
 import { convertFromUnits, isGreater } from "@/utils/bn";
 import BigNumber from "bignumber.js";
 import DateLib from "@/lib/date/DateLib";
-import { formatPercent } from "@/utils/formatter/percent";
 import { VotesSummaryHorizontalChart } from "@/src/modules/reporting/VotesSummaryHorizontalChart";
 import { useRetryUntilPassed } from "@/src/hooks/useRetryUntilPassed";
 import { t, Trans } from "@lingui/macro";
@@ -24,7 +23,7 @@ export const ActiveReportSummary = ({
   const startDate = DateLib.fromUnix(incidentReport.incidentDate);
   const endDate = DateLib.fromUnix(incidentReport.resolutionTimestamp);
 
-  const { formatCurrency } = useNumberFormat();
+  const { formatCurrency, formatPercent } = useNumberFormat();
 
   const votes = {
     yes: convertFromUnits(incidentReport.totalAttestedStake)
